HashMap: allow valueIterable to be iterated more than once
Iterators are stateful and can't be reused.  Therefore, the iterator protocol
makes Iterable's contain a zero-argument function which creates a new iterator,
and re-calls this zero-argument function to create a new iterator every time
an iteration starts.

hamt.values() returns an iterator which is stateful, so it must be wrapped in
a zero-argument function to become an Iterable<T>.