Switch custom node object inspection to use Symbol.for
util.inspect.custom is a symbol provided by node which allows custom
printing in the repl but is not available in the browser.  Bundlers
like webpack polyfill this by default but other bundlers like vite
do not (although vite can be configured to polyfill it).  But, to
avoid the need to polyfill at all, nodejs also exposes the symbol at
Symbol.for("nodejs.util.inspect.custom") on node v11 and up.

https://nodejs.org/api/util.html#util_util_inspect_custom

To avoid the need to polyfill, switch Value.ts to call
Symbol.for("nodejs.util.inspect.custom").  In node, this will use the
correct symbol to allow inspection and printing on the repl, while
in the browser it will create a new harmless symbol. As an added
benefit, the return type of Symbol.for is already a unique symbol so no
need for the explicit definition.

Note that Symbol.for is not available in IE11 and will need to be
polyfilled if you want to still support IE11.